feat: Highlight matches in 'Search'
