+++ client/package.json
@@ -70,7 +70,6 @@
     "chartjs-adapter-moment": "^1.0.1",
     "chroma-js": "^3.2.0",
     "debounce": "^3.0.0",
-    "dompurify": "^3.4.0",
     "jwt-decode": "^4.0.0",
     "moment": "^2.30.1",
     "oidc-client-ts": "^3.5.0",
